Prepare v3.3.0 release

+++ app/src/main/java/online/taxcore/pos/extensions/DoubleExtension.kt
@@ -11,7 +11,7 @@ fun Double.roundToDecimal(places: Int = 2) =
     BigDecimal(this).setScale(places, RoundingMode.HALF_UP).toString()
 
 fun Double.roundingToDecimal(places: Int) =
-    BigDecimal(this).setScale(places, RoundingMode.HALF_UP).toFloat()
+    BigDecimal(this).setScale(places, RoundingMode.HALF_UP).toDouble()
 
 val File.size get() = if (!exists()) 0.0 else length().toDouble()
 val File.sizeInKb get() = size / 1024
